watermark-detection/requirements.txt
torch
torchvision
pillow
numpy
matplotlib
tqdm
huggingface-hub
opencv-python
timm>=0.6.12
pandas==2.2.2
scikit_learn==1.5.0